Improve references file handling
Authors are now stored in the "authors" pubprop rather than
"pubmed_authors", abstract now in "abstract" rather than "pubmed_abstract".
We do this so that the authors and abstract sources are clear in Chado.

Parsing is now slightly more strict and the ID can be "pb_ref_id" or
"ref_id".

Refs geneontology/go-site#1762